backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain-google-genai
neo4j
pinecone
pydantic
supabase
langgraph
tavily-python
langchain-community
sqlalchemy
pgvector
alembic
psycopg2-binary
python-multipart
pyjwt
passlib
apscheduler
httpx
